refactor(step1): 콘솔 전용 파일을 step1 디렉터리로 이동
