+++ __tests__/validation/commonTest.js
@@ -1,5 +1,5 @@
 import { ERROR_MESSAGE } from "../../src/constants/message";
-import { isRetryAnswerValid } from "../../src/validates/RetryAnswerValidator";
+import { isRetryAnswerValid } from "../../src/step1/validates/RetryAnswerValidator";
 
 describe("프로그램이 끝났을 때 재시도 입력(y/n) 유효성 검증", () => {
   test("y를 입력하면 true를 반환한다.", () => {
