+++ src/step1-index.js
@@ -1,4 +1,4 @@
-import App from "./step1-App.js";
+import App from "./step1/step1-App.js";
 
 const app = new App();
 await app.run();
