+++ src/step1/views/InputView.js
