+++ src/step1/views/OutputView.js
@@ -1,5 +1,5 @@
 import { Console } from "@woowacourse/mission-utils";
-import { OUTPUT_MESSAGE } from "../constants/message.js";
+import { OUTPUT_MESSAGE } from "../../constants/message.js";
 import { formatNumber } from "../utils/FormatNumber.js";
 
 const OutputView = {
